requirements.txt
pandas
numpy
scikit-learn
matplotlib
seaborn
torch
optuna
pyyaml
boto3